analytics-py/requirements.txt
fastapi==0.115.4
uvicorn==0.30.6
httpx==0.27.2
pydantic==2.9.2
openai==1.55.3
python-dotenv==1.0.1
pandas==2.2.3
numpy==2.1.3
feedparser==6.0.11
cachetools==5.5.0
redis==5.1.1
PyYAML==6.0.2
psycopg[binary]==3.2.3
